+++ algorithms/models/Employees.py
@@ -12,19 +12,19 @@ employees: [Model] = [
     Model(
         key="2013310016",
         type="value",
-        text="gusion persley 2013310016",
-        vector="gusion persley 2013310016"
+        text="gusion persley",
+        vector="gusion persley"
     ),
     Model(
         key="2013310017",
         type="value",
-        text="aamon persley 2013310017",
-        vector="aamon persley 2013310017"
+        text="aamon persley",
+        vector="aamon persley"
     ),
     Model(
         key="2013310018",
         type="value",
-        text="valentina persley 2013310018",
-        vector="valentina persley 2013310018"
+        text="valentina persley",
+        vector="valentina persley"
     ),
 ]